addMessageServeDelete，addMessageServeConsent，getFriendInfo，完成好友添加请求，好友添加，好友列表，后端代码总行数： 1560

+++ src/main/java/com/example/chatserver/bean/Friendlies.java
@@ -17,6 +17,6 @@ import lombok.ToString;
 @Data
 public class Friendlies {
     private Long _id;
-    private Long userX;
-    private Long userY;
+    private int userX;
+    private int userY;
 }
